+++ src/components/grid/grid.js
@@ -50,6 +50,16 @@ import { startCase }  from 'lodash';
  *                    {name: 'credit', align: 'right'},
  *                    {name: 'total', align: 'right'}];
  *
+ * To add a deleting action to the grid pass you can pass an additional callback prop
+ *
+ *    onRowDelete - callback when delete action is triggered
+ *
+ *    <Grid
+ *      fields={ fields }
+ *      data={ data.get('line_items') }
+ *      onRowClick={ this.handleRowClick }
+ *      onRowDelete={ this.handleRowDelete } />
+ *
  * @class Grid
  * @extends React.Component
  */
@@ -81,7 +91,26 @@ class Grid extends React.Component {
      * @property onRowClick
      * @type {Function}
      */
-    onRowClick: React.PropTypes.func
+    onRowClick: React.PropTypes.func,
+
+    /**
+     * A callback for when a row delete action is triggered.
+     *
+     * @property onRowDelete
+     * @type {Function}
+     */
+    onRowDelete: React.PropTypes.func
+  }
+
+  /**
+   * Builds a placeholder column for the delete action
+   *
+   * @method deletingColumn
+   * @return {Object} JSX
+   */
+  get deletingColumn() {
+    let className = this.cellClasses({}) + 'ui-grid__header__cell--delete';
+    return ( <th key='delete' className={ className }></th> );
   }
 
   /**
@@ -91,13 +120,21 @@ class Grid extends React.Component {
    * @return {Object} JSX
    */
   get columns() {
-    return this.props.fields.map((column, index) => {
+    let columns = [];
+
+    if (this.props.onRowDelete) {
+      columns.push(this.deletingColumn);
+    }
+
+    this.props.fields.forEach((column, index) => {
       let displayName = column.displayName || startCase(column.name);
 
-      return (
+      columns.push(
         <th key={ index } className={ this.cellClasses(column) }>{ displayName }</th>
       );
     });
+
+    return columns;
   }
 
   /**
@@ -107,13 +144,17 @@ class Grid extends React.Component {
    * @return {Object} JSX
    */
   get rows() {
-    return this.props.data.map((row, index) => {
+    return this.props.data.map((row) => {
+      let row_id = row.get('_row_id');
+
       return (
         <GridRow
-          key={index}
+          key={ row_id }
           fields={ this.props.fields }
-          row={ row }
+          data={ row }
+          row_id={ row_id }
           onRowClick={ this.props.onRowClick }
+          onRowDelete={ this.props.onRowDelete }
         />
       );
     });