+++ example/views/dummy/index.js
@@ -7,6 +7,7 @@ import Link from   'components/link';
 
 import FinancesStore from './../../stores/finances';
 import UserActions from './../../actions/user';
+import FinancesActions from './../../actions/finances';
 
 import FinancesHistory from './subviews/history';
 import FinancesDetails from './subviews/details';
@@ -16,9 +17,22 @@ import UserDialog from './subviews/user-dialog';
 
 
 class Finances extends React.Component {
+
+  componentWillUpdate(nextProps, nextState) {
+    if (this.state.financesStore.get('success')) {
+      console.log('saved!');
+    }
+  }
+
   handleOnClick = (ev) => {
     ev.preventDefault();
     UserActions.userDialogOpened();
+  }
+
+  save = (ev) => {
+    ev.preventDefault();
+    FinancesActions.beforeSave();
+    FinancesActions.financesSave();
   }
 
   render() {
@@ -34,7 +48,7 @@ class Finances extends React.Component {
 
         <h1 className="view-finances__title">{ name }</h1>
 
-        <Form model="foo">
+        <Form model="foo" saving={ this.state.financesStore.get('isSaving') } afterFormValidation={ this.save } ref="form">
           <FinancesDetails
             name={ name }
             countryValue={ this.state.financesStore.getIn(['country', 'id']) }