+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 * Grid
 * Animated Menu Button
 
+* Input Grid - Breaking change
+
+Input Grid now requires a onRowDelete prop and onRowUpdate prop instead of a deleteRowHandler prop and updateRowHandler prop
+
 ## Minor
 
 * Input Grid is now decorated by common grid